+++ src/pages/Friend/Friend.module.css
@@ -59,7 +59,7 @@
 
 .recommend {
     display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(550px, 1fr));
+    grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
     grid-row-gap: 20px;
     justify-content: center;
     gap: 30px;
