+++ src/pages/Friend/components/RecommendFriendCard.module.css
@@ -1,0 +1,59 @@
+.card {
+    display: inline-grid;
+    color: white;
+    transition: transform 0.3s; 
+    transform: perspective(800px) rotateY(0deg);
+    transform-style: preserve-3d;
+
+    width: 200px;
+    height: 350px;
+}
+
+.card > * {
+    grid-area: 1 / 1 / 1 / 1;
+    padding: 5px;
+    border-radius: 8px;
+    backface-visibility: hidden;
+}
+
+.front {
+    width: 200px;
+    height: 350px;
+    background: #C8DCA0 
+}
+
+.back {
+    background: linear-gradient(rgb(127, 106, 83), rgb(156, 132, 104));
+    transform: rotateY(180deg);
+}
+
+/* .profileimg{
+    width: 80%;
+    border-radius: 50%;
+} */
+
+.content {
+    background-color: #FFFFFF;
+    border-radius: 8px;
+    width: 100%;
+    height: 100%;
+    text-align: center;
+    padding: 10px 5px;
+}
+
+.flipped {
+    transform: perspective(800px) rotateY(180deg);
+}
+
+.introduce {
+    line-height: 1.2em;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    font-size: 12px;
+    text-align: left;
+    white-space: normal;
+    word-wrap: break-word;
+    display: -webkit-box;
+    -webkit-line-clamp: 2;
+    -webkit-box-orient: vertical;
+}
